requirements.txt
requests==2.31.0
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
httpx==0.24.1
openai==1.14.0
beautifulsoup4>=4.9.3,<5.0.0
msal~=1.20
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pyodbc>=4.0.39
gunicorn>=21.2.0

